Back/requirements.txt
django==4.2
psycopg2-binary
djangorestframework
django-cors-headers
pillow
